Move logic to build keyword matcher into apk file and use singleton
